--- client/app/case/case.module.js
@@ -1,11 +1,0 @@
-/*
-Created by: Prado Arturo Bognot
-Position: SWAT Supervising SRS
-Date: Sept 6, 2017
-*/
-
-(function() {
-    'use strict';
-
-    angular.module('app.case', []);
-})();
