--- client/app/case/testCase.html
@@ -1,119 +1,0 @@
-<div class="container">
-	<div class="row">
-		<div class="col-sm-10">
-			<div class="form-group">
-				<label>Test Case Name: <small>(<a href="#" ng-click="vm.showData()">Show Data</a>)</small></label>
-				<input type="text" class="form-control" value="Enter test case name..." ng-model="vm.test_case">
-			</div>
-		</div>
-		<div class="col-sm-2">
-			<div class="form-group">
-				<label>Global Wait (ms):</label>
-				<input type="number" class="form-control" ng-model="vm.global_wait">
-			</div>
-		</div>
-	</div>
-</div>
-<div dnd-list="vm.case_steps">
-	<div ng-repeat="indiv_step in vm.case_steps"
-         dnd-draggable="indiv_step"
-         dnd-moved="vm.case_steps.splice($index, 1)"
-         dnd-effect-allowed="move"
-         dnd-selected="vm.selected = indiv_step"
-         ng-class="{'selected': vm.selected === indiv_step}">
-		<div class="case-step">
-		    <div class="step">
-		        <div class="item">
-		            {{$index + 1}}.
-		        </div>
-		        <div class="item">
-		            <div class="form-group">
-		                <label>Step Name:</label>
-		                <input type="text" class="form-control" ng-model="indiv_step.name">
-		            </div>
-		        </div>
-		        <div class="item">
-		            <div class="form-group">
-		                <label>Action:</label>
-		                <select class="form-control" ng-model="indiv_step.action" ng-change="vm.selectedAction(indiv_step)">
-		                    <option>Choose action</option>
-		                    <option>Click</option>
-		                    <option>Hover</option>
-		                    <option>Write Text</option>
-		                    <option>Go to URL</option>
-		                    <option>Select Option</option>
-		                    <option>Select Random Option</option>
-		                    <option>Wait</option>
-		                    <option>Add Assertion</option>
-		                    <option>Maximize Window</option>
-		                    <option>Modify Window Size</option>
-		                    <!-- <option>Close Alert Box</option> -->
-		                </select>
-		            </div>
-		        </div>
-		        <div class="item" ng-show="indiv_step.assert_options.show">
-		        	<div class="form-group">
-		        		<label>Assertion Type:</label>
-		                <select class="form-control" ng-model="indiv_step.assert_options.type" ng-change="vm.selectedAssertion(indiv_step)">
-		                	<option>Choose Type</option>
-		                    <option>Element is Visible</option>
-		                    <option>Element is Not Visible</option>
-		                    <option>Matches Value / RegEx</option>
-		                    <option>Does Not Match Value / RegEx</option>
-		                    <option>Contains Value</option>
-		                    <option>Does Not Contain Value</option>
-		                    <option>Title Contains Value</option>
-		                    <option>URL Contains Value</option>
-		                </select>
-		        	</div>
-		        </div>
-		        <div class="item" ng-show="indiv_step.locateElement.show">
-		        	<div class="form-group">
-		        		<label>Locate Element By:</label>
-		                <select class="form-control" ng-model="indiv_step.locateElement.by" ng-change="vm.selectedLocator(indiv_step)">
-		                    <option>Choose locator</option>
-		                    <option>Id</option>
-		                    <!-- <option>Text Inside</option> -->
-		                    <option>Name</option>
-		                    <option>Class Name</option>
-		                    <option>XPath</option>
-		                    <option>CSS Selector</option>
-		                    <option>Link Text</option>
-		                    <option>Partial Link Text</option>
-		                </select>
-		        	</div>
-		        </div>
-		        <div class="item-smaller" ng-show="indiv_step.element_id.show">
-		        	<div class="form-group">
-		        		<label>{{indiv_step.locateElement.by}}:</label>
-		            <input type="text" class="form-control" ng-model="indiv_step.element_id.value">
-		        	</div>
-		        </div>
-		        <div class="item-smaller" ng-show="indiv_step.op_special_1.show">
-		        	<div class="form-group">
-		        		<label>{{indiv_step.op_special_1.label}}:</label>
-		            <input type="text" class="form-control" ng-model="indiv_step.op_special_1.value">
-		        	</div>
-		        </div>
-		        <div class="item-smaller" ng-show="indiv_step.op_special_2.show">
-		        	<div class="form-group">
-		        		<label>{{indiv_step.op_special_2.label}}:</label>
-		            <input type="text" class="form-control" ng-model="indiv_step.op_special_2.value">
-		        	</div>
-		        </div>
-		    </div>
-		    <div class="beats-action">
-		        <div class="item">
-		            <a href="#" ng-click="vm.removeStep($index)">Delete</a>
-		        </div>
-		    </div>
-		</div>
-		<hr>
-	</div>
-</div>
-
-<div class="test-case-bottom-buttons">
-    <button class="btn btn-default" ng-click="vm.addStep()">Add step</button>
-    <!-- <button download="testScript.txt" class="btn btn-warning" ng-click="vm.downloadTestCase()">Download Test Script</button> -->
-    <a href="#" class="btn btn-warning" role="button" ng-click="vm.downloadTestCase()">Download Test Script</a>
-</div>
