--- client/app/hello.htm
@@ -1,2 +1,0 @@
-<h3> {{ $ctrl.greeting }} solar system! </h3>
-<button ng-click="$ctrl.toggleGreeting()">toggle greeting</button>
