--- client/app/hello.js
@@ -1,11 +1,0 @@
-angular.module('beatsApp').component('hello', {
-  templateUrl: '/client/app/hello.htm',
-
-  controller: function() {
-    this.greeting = 'hello';
-
-    this.toggleGreeting = function() {
-      this.greeting = (this.greeting == 'hello') ? 'whats up' : 'hello'
-    }
-  }
-});
