--- client/app/home/home.htm
@@ -1,8 +1,0 @@
-<!-- 
-Created by: Prado Arturo Bognot
-Position: SWAT Supervising SRS
-Date: Oct 6, 2017 
--->
-
-<h3> {{ $ctrl.greeting }} solar system! </h3>
-<button ng-click="$ctrl.toggleGreeting()">toggle greeting</button>
