+++ client/app/services/downloadService.js
