--- client/app/step/caseStep.html
@@ -1,37 +1,0 @@
-<div class="case-step">
-    <div class="step">
-        <div class="item">
-            {{step.id}}.
-        </div>
-        <div class="item">
-            <div class="form-group">
-                <label>Step Name:</label>
-                <input type="text" class="form-control">
-            </div>
-        </div>
-        <div class="item">
-            <div class="form-group">
-                <label>Action:</label>
-                <select class="form-control">
-                    <option>Choose action</option>
-                    <option>Click</option>
-                    <option>Write Text</option>
-                    <option>Go to URL</option>
-                    <option>Select Option</option>
-                    <option>Add Assertion</option>
-                    <option>Import Test Case</option>
-                    <option>Close Alert Box</option>
-                </select>
-            </div>
-        </div>
-    </div>
-    <div class="beats-action">
-        <div class="item">
-            <a href="#">Move</a>
-        </div>
-        <div class="item">
-            <a href="#">Delete</a>
-        </div>
-    </div>
-</div>
-<hr>
